feat: enhance button functionality and update instruction handling across components
- Added ShareButton to molecules and integrated it into the Export page.
- Removed ZoomButton and DownloadButton components, streamlining button options.
- Replaced Instructions component with InstructionsModal in Import and Edit pages for improved modal handling.
- Updated organism exports to reflect new modal structure and removed deprecated components.
- Introduced InstructionMode type for better instruction management.
