+++ src/components/organisms/Instructions/AlwaysShowToggleSwitch.tsx
@@ -1,0 +1,14 @@
+import { SettingsToggle } from '@molecules';
+import { usePersistedSettingsStore } from '@stores';
+
+export const AlwaysShowToggleSwitch: React.FC = () => {
+  const { alwaysShowInstructions, setAlwaysShowInstructions } =
+    usePersistedSettingsStore();
+  return (
+    <SettingsToggle
+      title="show every time"
+      isEnabled={alwaysShowInstructions}
+      onToggle={setAlwaysShowInstructions}
+    />
+  );
+};
