+++ src/components/organisms/index.ts
@@ -1,4 +1,4 @@
 export { NavigationBar } from './NavigationBar';
 export { SelectionOverlay } from './SelectionOverlay';
 export { PanZoomControl } from './PanZoomControl';
-export { Instructions } from './Instructions';
+export { InstructionsModal } from './Instructions/InstructionsModal';
