+++ src/components/pages/export/Export.tsx
@@ -2,17 +2,27 @@ import React from 'react';
 import { useTransformedImageStore } from '@stores';
 import { Image } from 'react-native';
 import { PageTemplate } from '@templates';
-import { DownloadButton } from '@molecules';
+import { DownloadButton, ShareButton } from '@molecules';
+import { useScreenDimensions } from '@hooks';
+import { InstructionsModal } from '@organisms';
+import { InstructionMode } from '@types';
 
 export const Export: React.FC = () => {
   const { destinationUri } = useTransformedImageStore();
-
+  const { isMobile } = useScreenDimensions();
   if (!destinationUri) return null;
 
   return (
     <PageTemplate>
+      <PageTemplate.ModalContent>
+        <InstructionsModal mode={InstructionMode.EXPORT} />
+      </PageTemplate.ModalContent>
       <PageTemplate.ActionItems>
-        <DownloadButton key="download-button" />
+        {isMobile ? (
+          <ShareButton key="share-button" />
+        ) : (
+          <DownloadButton key="download-button" />
+        )}
       </PageTemplate.ActionItems>
       <Image
         source={{ uri: destinationUri }}
