+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export { useTransformImage } from './useTransformImage';
 export { useShareImage } from './useShareImage';
 export { usePageTemplateContext } from './usePageTemplateContext';
 export { usePageModalContext } from './usePageModalContext';
+export { useAutoShowInstructions } from './useAutoShowInstructions';
